Upgrade to Scala 2.13.2 and ScalaJS 1.1.0

+++ client/src/main/scala/moleculeadmin/client/app/logic/query/keyEvents/Paging.scala
@@ -49,7 +49,7 @@ trait Paging extends BaseKeyEvents {
       t2 = js.Date.now
       delta = t2 - t1
       tProcess = tProcess + delta
-      avg = (tProcess / j).round
+      avg = (tProcess / j).toDouble.round
       ratio = avg / keyRepeatMs
       cycles = ratio.ceil.toInt
       // println(s"  $j  $delta    $avg    $ratio    $cycles    " + (t2 - beginning))
